# agent/design/api-contract.yaml
openapi: 3.0.3
info:
  title: PumpBuddy API
  version: 0.1.0
  description: |
    Canonical API contract for the active workout product slice.
    Note: the 2026-03-28 auth/session boundary refactor is internal-only and does
    not change external endpoints or payload schemas.
    Note: the 2026-04-06 workout-select greeting update reuses existing
    AuthSessionUser.display_name and does not change endpoint shapes.
    Note: the 2026-04-12 auth-failure logging hardening emits structured
    `auth_login_failure` telemetry for application attempt keys without
    endpoint or payload-schema changes.
    Note: the 2026-04-12 settings profile data update extends AuthSessionUser
    with `login` and `registration_date` for settings display.
    Note: the 2026-04-12 settings display-name persistence update adds PATCH
    /auth/session for authenticated display-name changes.
    Note: the 2026-04-13 settings preferences contract update extends PATCH
    /auth/session and AuthSessionUser with `favorite_gym_id`, and adds
    POST /auth/password for authenticated password changes.
    Note: the 2026-05-30 password-change lifecycle hardening keeps
    /auth/password request and response schemas unchanged while throttling
    repeated current-password failures and clearing the session cookie after a
    successful password change.
    Note: the 2026-04-14 About navigation wiring tightened renderer
    AboutMetadata.channel typing to `stable`; this is API-surface alignment
...
paths:
...
  /api/gyms/{gymId}:
    get:
      summary: Retrieve read-only gym transparency detail
      operationId: getGymDetail
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Gym detail retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GymDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Gym not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      summary: Rename a gym
      operationId: updateGym
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GymUpdateRequest'
      responses:
        '200':
          description: Gym renamed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GymSummary'
        '400':
          description: Invalid gym update payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Gym not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate normalized gym name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete a draft gym
      operationId: deleteGym
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Draft gym deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Gym not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Gym cannot be deleted in its current state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  responses:
    Unauthorized:
      description: Unauthorized (empty response body)
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    GymSummary:
      type: object
      required:
        - id
        - name
        - status
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [new, active, inactive]
        station_count:
          type: integer
          format: int64
          minimum: 0
          nullable: true
          description: |
            Count of stations visible for this gym. Optional during staged
            rollout so existing start-screen gym consumption remains compatible.
        last_visited_at:
          type: string
          format: date-time
          nullable: true
          description: |
            Most recent completed workout timestamp for this gym.
            Null when the authenticated user has no completed visits.
            Optional during staged rollout so existing start-screen gym
            consumption remains compatible.
    GymDetailResponse:
      type: object
      description: |
        Read-only gym transparency payload for renderer detail browsing.
        Ordering of stations, exercise groups, variants, and station options is
        backend-owned and deterministic by display name with stable id
        tie-breakers.
      required:
        - id
        - name
        - status
        - station_count
        - last_visited_at
        - stations
        - exercise_groups
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [new, active, inactive]
        station_count:
          type: integer
          format: int64
          minimum: 0
        last_visited_at:
          type: string
          format: date-time
          nullable: true
          description: |
            Most recent completed workout timestamp for this gym.
            Null when the authenticated user has no completed visits.
        stations:
          type: array
          items:
            $ref: '#/components/schemas/GymStationSummary'
        exercise_groups:
          type: array
          description: Exercise variants grouped by canonical exercise.
          items:
            $ref: '#/components/schemas/GymExerciseGroup'
...
    GymUpdateRequest:
      type: object
      additionalProperties: false
      description: Name-only payload. Lifecycle status is server-owned and must not be supplied.
      required:
        - name
      properties:
        name:
          type: string
    GymStationSummary:
      type: object
      required:
        - id
        - name
        - load_profile_name
        - suitable_variant_count
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        load_profile_name:
          type: string
          description: Display name of the station load profile.
        suitable_variant_count:
          type: integer
          format: int64
          minimum: 0
          description: Number of station-required variants compatible with this station.
...
    GymExerciseGroup:
      type: object
      required:
        - exercise_id
        - exercise_name
        - variants
      properties:
        exercise_id:
          type: string
          format: uuid
        exercise_name:
          type: string
        variants:
          type: array
          items:
            $ref: '#/components/schemas/GymExerciseVariantSummary'
    GymExerciseVariantSummary:
      type: object
      description: |
        Gym transparency variant projection. Stationless rows are represented by
        requires_station=false, station_availability=STATIONLESS, and an empty
        station_options array rather than a synthetic station record.
      required:
        - variant_id
        - variant_name
        - requires_station
        - station_availability
        - repetition_kind
        - load_input_mode
        - set_tracking_mode
        - station_options
      properties:
        variant_id:
          type: string
          format: uuid
        variant_name:
          type: string
        requires_station:
          type: boolean
          description: Canonical ExerciseVariant.requires_station value.
        station_availability:
          type: string
          enum: [STATIONLESS, SINGLE_STATION, MULTI_STATION]
          description: |
            Backend-derived station suitability bucket for renderer navigation.
            STATIONLESS opens exercise variant detail directly.
            SINGLE_STATION opens that station detail directly.
            MULTI_STATION requires station selection before station detail.
        repetition_kind:
          type: string
          enum: [REPS, SECS]
          description: Variant repetition unit used by renderer detail copy.
        load_input_mode:
          type: string
          enum: [TOTAL, PER_SIDE]
          description: Variant load entry mode.
        set_tracking_mode:
          type: string
          enum: [UNILATERAL, BILATERAL]
          description: Variant set-tracking mode.
        station_options:
          type: array
          description: |
            Compatible station options for station-required variants in this
            gym. Empty for stationless variants.
          items:
            $ref: '#/components/schemas/GymStationOption'
    GymStationOption:
      type: object
      required:
        - station_id
        - station_name
      properties:
        station_id:
          type: string
          format: uuid
        station_name:
          type: string
        station_profile_loads_kg:
          type: array
          description: |
            Optional station load-profile values in canonical kg. Present when
            a caller needs station range display, such as training-plan detail
            compatible-station rows. Omitted for compact gym station options.
          items:
            type: number
            format: double
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        details:
          $ref: '#/components/schemas/ErrorDetails'
    ErrorDetails:
      type: object
      properties:
        selected_gym_id:
          type: string
          format: uuid
          description: |
            Gym selected for configured-gym workout start validation.
            Returned when workout start is blocked because one or more exercises have no realizable option in that gym.
            Equals the configured-gym context used for start validation.
        missing_exercises:
          type: array
          description: |
            Plan exercises that block configured-gym workout start in the selected gym context.
            Each item identifies the training-plan exercise that has no realizable option after compatibility filtering.
            Items are sorted by exercise_position ascending, then training_plan_exercise_id.
          items:
            $ref: '#/components/schemas/MissingExerciseDetail'
    MissingExerciseDetail:
      type: object
      required:
        - training_plan_exercise_id
        - exercise_name
        - exercise_position
        - reason
      properties:
        training_plan_exercise_id:
          type: string
          format: uuid
        exercise_name:
          type: string
        exercise_position:
          type: integer
          format: int32
          minimum: 1
        reason:
          type: string
...